// src/theme/types.ts
export interface ThemeColors {
    background: string;
    backgroundGradient: string[];
    surface: string;
    card: string;

    text: string;
    textSecondary: string;
    textTertiary: string;

    border: string;
    borderLight: string;

    inputBackground: string;
    inputBorder: string;
    inputText: string;
    placeholder: string;

    primary: string;
    primaryText: string;

    success: string;
    warning: string;
    error: string;
    info: string;

    overlay: string;
    shadow: string;
    badge: string;

    icon: string;
    iconSecondary: string;
    badgePrimary: string;
    weekDateBackground: string;
    weekText: string;
    timeCardBackground: string,
    timeText: string,
    timeFeather: string
}